+++ examples/react-native-todo-list/assets/svg/ConfirmIcon.js
@@ -1,5 +1,5 @@
-import * as React from 'react';
-import { SvgXml } from 'react-native-svg';
+import * as React from "react";
+import { SvgXml } from "react-native-svg";
 
 const activeIcon = `
 <svg width="36" height="36" viewBox="0 0 36 36" fill="none" xmlns="http://www.w3.org/2000/svg">