+++ examples/react-native-todo-list/assets/svg/Delete.js
@@ -1,5 +1,5 @@
-import React from 'react';
-import { SvgXml } from 'react-native-svg';
+import React from "react";
+import { SvgXml } from "react-native-svg";
 
 const xml = ` 
 <svg width="28" height="28" viewBox="0 0 28 28" fill="none" xmlns="http://www.w3.org/2000/svg">