+++ examples/react-native-todo-list/metro.config.js
@@ -1,23 +1,16 @@
-/**
- * Metro configuration for React Native
- * https://github.com/facebook/react-native
- *
- * @format
- */
+const { getDefaultConfig } = require("metro-config");
 
- const { getDefaultConfig } = require('metro-config');
-
- module.exports = (async () => {
-   const {
-     resolver: { sourceExts, assetExts },
-   } = await getDefaultConfig();
-   return {
-     transformer: {
-       babelTransformerPath: require.resolve('react-native-svg-transformer'),
-     },
-     resolver: {
-       assetExts: assetExts.filter(ext => ext !== 'svg'),
-       sourceExts: [...sourceExts, 'svg'],
-     },
-   };
- })();
+module.exports = (async () => {
+  const {
+    resolver: { sourceExts, assetExts },
+  } = await getDefaultConfig();
+  return {
+    transformer: {
+      babelTransformerPath: require.resolve("react-native-svg-transformer"),
+    },
+    resolver: {
+      assetExts: assetExts.filter((ext) => ext !== "svg"),
+      sourceExts: [...sourceExts, "svg"],
+    },
+  };
+})();